pyproject.toml
[tool.poetry]
name = "Lincride Test"
version = "0.1.0"
description = ""
authors = ["dakohhh <wiizzydreadmill@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-environ = "^0.11.2"
dj-database-url = "^2.3.0"
psycopg2-binary = "^2.9.10"
django-mjml = "^1.3"
drf-yasg = "^1.21.8"
drf-spectacular = "^0.28.0"
gunicorn = "^23.0.0"
uvicorn = "^0.34.0"
uvicorn-worker = "^0.3.0"
django-cors-headers = "^4.6.0"
pendulum = "^3.0.0"


[tool.poetry.group.dev.dependencies]
black = "24.8.0"
ruff = "^0.8.3"
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''
skip-string-normalization = true
skip-magic-trailing-comma = true
verbose = true
fast = true
